+++ src/f2_helper/f2_helper.h
@@ -15,7 +15,7 @@ int sign_f2_dot_product(int x, int y);
 
 /// Gives (i)^(x.y), where . is the F_2 inner product between 2
 /// F_2 vectors (represented as integers)
-std::complex<int> imag_f2_dot_product(int x, int y);
+std::complex<float> imag_f2_dot_product(int x, int y);
 
 }
