+++ src/f2_helper/f2_helper.test.cpp
@@ -28,5 +28,5 @@ TEST_CASE("imaginary mod 2 product", "[f2 helper]") {
     int z = 45; // 101101
 
     REQUIRE(imag_f2_dot_product(x,y) == 1);
-    REQUIRE(imag_f2_dot_product(x,z) == std::complex<int> (0, 1));
+    REQUIRE(imag_f2_dot_product(x,z) == std::complex<float> (0, 1));
 }
